requirements.txt
python-telegram-bot==20.6
openai==1.14.3
